CedFRA_openc/app/requirements.txt
python-docx>=1.1.0
pdfplumber>=0.11.0
num2words>=0.5.13
holidays>=0.63
pytest>=8.0.0
